+++ src/lib.rs
@@ -103,6 +103,11 @@ pub trait Filesystem {
 		Err(Error::from_raw_os_error(libc::ENOSYS))
 	}
 
+	fn create(&mut self, _req: &Request, path: &Path, mode: u32, info: &FileInfo) -> Result<()> {
+		let _ = (path, mode, info);
+		Err(Error::from_raw_os_error(libc::ENOSYS))
+	}
+
 	fn chown(&mut self, _req: &Request, path: &Path, uid: Option<u32>, gid: Option<u32>) -> Result<()> {
 		let _ = (path, uid, gid);
 		Err(Error::from_raw_os_error(libc::ENOSYS))