+++ ChangeLog.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Filesystem::symlink()
 - Filesystem::rename()
 - Filesystem::truncate()
+- Filesystem::create()
 
 ## [0.0.3] - 2024-10-25